+++ av/src/VideoEncoder.cc
@@ -15,10 +15,6 @@
  *
 */
 #include <stdio.h>
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <fcntl.h>
-#include <sys/ioctl.h>
 
 #include <mutex>