+++ graphics/src/SkeletonAnimation.cc
@@ -69,6 +69,16 @@ unsigned int SkeletonAnimation::NodeCount() const
 }
 
 //////////////////////////////////////////////////
+NodeAnimation *SkeletonAnimation::NodeAnimationByName(
+    const std::string &_node) const
+{
+  auto it = this->data->animations.find(_node);
+  if (it != this->data->animations.end())
+    return it->second;
+  return nullptr;
+}
+
+//////////////////////////////////////////////////
 bool SkeletonAnimation::HasNode(const std::string &_node) const
 {
   return (this->data->animations.find(_node) != this->data->animations.end());
@@ -139,6 +149,12 @@ std::map<std::string, math::Matrix4d> SkeletonAnimation::PoseAtX(
 {
   std::map<std::string, NodeAnimation*>::const_iterator nodeAnim =
       this->data->animations.find(_node);
+  if (nodeAnim == this->data->animations.end())
+  {
+    ignerr << "Can't find animation named [" << _node << "]" << std::endl;
+    return {};
+  }
+
   math::Matrix4d lastPos = nodeAnim->second->KeyFrame(
       nodeAnim->second->FrameCount() - 1).second;