+++ include/ignition/common/Util.hh
@@ -25,6 +25,8 @@
 #include <chrono>
 #include <ignition/common/Export.hh>
 #include <ignition/common/Filesystem.hh>
+#include <ignition/common/SystemPaths.hh>
+#include <ignition/common/URI.hh>
 
 /////////////////////////////////////////////////
 // Defines
@@ -147,6 +149,25 @@ namespace ignition
     /// \param[in] _file the file name to look for.
     /// \return The path containing the file.
     std::string IGNITION_COMMON_VISIBLE findFilePath(const std::string &_file);
+
+    /// \brief Add a callback to use when findFile() can't find a file that is
+    /// a valid URI.
+    /// The callback should return a full local path to the requested file, or
+    /// and empty string if the file was not found in the callback.
+    /// Callbacks will be called in the order they were added until a path is
+    /// found.
+    /// \param[in] _cb The callback function, which takes a file path or URI
+    /// and returns the full local path.
+    void IGNITION_COMMON_VISIBLE addFindFileURICallback(
+        std::function<std::string(const URI &)> _cb);
+
+    /// \brief Get a pointer to the global system paths that is used by all
+    /// the findFile functions.
+    /// The returned instance has global shared state for a given process.
+    /// Care should be taken when manipulating global system paths
+    /// Caller should not assume ownership of the pointer.
+    /// \return A mutable reference to the system paths object.
+    common::SystemPaths IGNITION_COMMON_VISIBLE *systemPaths();
 
     /// \brief Compute the SHA1 hash of an array of bytes.
     /// \param[in] _buffer Input sequence. The permitted data types for this