+++ src/SystemPaths.cc
@@ -199,6 +199,12 @@ void SystemPaths::SetFilePathEnv(const std::string &_env)
 }
 
 /////////////////////////////////////////////////
+std::string SystemPaths::FilePathEnv() const
+{
+  return this->dataPtr->filePathEnv;
+}
+
+/////////////////////////////////////////////////
 const std::list<std::string> &SystemPaths::FilePaths()
 {
   return this->dataPtr->filePaths;
@@ -328,6 +334,21 @@ std::string SystemPaths::FindFileURI(const ignition::common::URI &_uri) const
     filename = this->dataPtr->findFileURICB(_uri.Str());
   }
 
+  // Look in custom paths.
+  // Tries the suffix against all paths, regardless of the scheme
+  if (filename.empty())
+  {
+    for (const std::string &filePath : this->dataPtr->filePaths)
+    {
+      auto withSuffix = NormalizeDirectoryPath(filePath) + suffix;
+      if (exists(withSuffix))
+      {
+        filename = ignition::common::copyFromUnixPath(withSuffix);
+        break;
+      }
+    }
+  }
+
   // If still not found, try custom callbacks
   if (filename.empty())
   {
@@ -408,10 +429,10 @@ std::string SystemPaths::FindFile(const std::string &_filename,
   {
     for (const std::string &filePath : this->dataPtr->filePaths)
     {
-      if (exists(NormalizeDirectoryPath(filePath) + filename))
+      auto withSuffix = NormalizeDirectoryPath(filePath) + filename;
+      if (exists(withSuffix))
       {
-        path = NormalizeDirectoryPath(filePath) + filename;
-        ignition::common::replaceAll(path, path, "//", "/");
+        path = ignition::common::copyFromUnixPath(withSuffix);
         break;
       }
     }