+++ src/SystemPaths_TEST.cc
@@ -34,6 +34,7 @@ using namespace ignition;
 
 class SystemPathsFixture : public ::testing::Test
 {
+  // Documentation inherited
   public: virtual void SetUp()
     {
       this->backupPluginPath = "IGN_PLUGIN_PATH=";
@@ -57,14 +58,20 @@ class SystemPathsFixture : public ::testing::Test
 #endif
     }
 
+  // Documentation inherited
   public: virtual void TearDown()
     {
       putenv(const_cast<char*>(this->backupPluginPath.c_str()));
       putenv(const_cast<char*>(this->backupFilePath.c_str()));
     }
 
+  /// \brief Backup of plugin paths to be restored after the test
   public: std::string backupPluginPath;
+
+  /// \brief Backup of file paths to be restored after the test
   public: std::string backupFilePath;
+
+  /// \brief Root of filesystem according to each platform
   public: std::string filesystemRoot;
 };
 
@@ -108,6 +115,7 @@ TEST_F(SystemPathsFixture, FileSystemPaths)
   putenv(env);
 
   common::SystemPaths paths;
+  EXPECT_EQ("IGN_FILE_PATH", paths.FilePathEnv());
   const std::list<std::string> pathList3 = paths.FilePaths();
   EXPECT_EQ(static_cast<unsigned int>(2), pathList3.size());
   EXPECT_STREQ("/tmp/file/", pathList3.front().c_str());
@@ -119,8 +127,8 @@ TEST_F(SystemPathsFixture, FileSystemPaths)
 
   std::string dir1 = "test_dir1";
   ignition::common::createDirectories(dir1);
-  std::string file1 =
-      ignition::common::SystemPaths::NormalizeDirectoryPath(dir1) + "test_f1";
+  auto file1 = ignition::common::copyFromUnixPath(
+      ignition::common::joinPaths(dir1, "test_f1"));
   std::ofstream fout;
   fout.open(file1, std::ofstream::out);
   fout << "asdf";
@@ -135,6 +143,7 @@ TEST_F(SystemPathsFixture, FileSystemPaths)
   putenv(env);
   paths.SetFilePathEnv("IGN_FILE_PATH");
   EXPECT_EQ(file1, paths.FindFile("test_f1")) << paths.FindFile("test_f1");
+  EXPECT_EQ(file1, paths.FindFile("model://test_f1"));
 }
 
 /////////////////////////////////////////////////
@@ -213,7 +222,7 @@ TEST_F(SystemPathsFixture, FindFileURI)
           this->filesystemRoot);
 
   EXPECT_EQ("", sp.FindFileURI("file://no_such_file"));
-  EXPECT_EQ(file1, sp.FindFileURI("file:test_dir1/test_f1"));
+  EXPECT_EQ(file1, sp.FindFileURI("file://test_dir1/test_f1"));
   EXPECT_EQ(file1, sp.FindFileURI("file://" +
                                   ignition::common::copyToUnixPath(file1)));
   EXPECT_EQ("", sp.FindFileURI("osrf://unknown.protocol"));
@@ -267,13 +276,32 @@ TEST_F(SystemPathsFixture, FindFileURI)
 
   sp.AddFindFileURICallback(osrfCb);
   EXPECT_EQ(file1, sp.FindFileURI("robot://test_f1"));
-  EXPECT_EQ(file2, sp.FindFileURI("osrf:test_f2"));
+  EXPECT_EQ(file2, sp.FindFileURI("osrf://test_f2"));
 
   // Test that th CB from SetFindFileURICallback is called first even when a
   // second handler for the same protocol is available
   sp.AddFindFileURICallback(robot2Cb);
   EXPECT_EQ(file1, sp.FindFileURI("robot://test_f1"));
-  EXPECT_EQ(file2, sp.FindFileURI("osrf:test_f2"));
+  EXPECT_EQ(file2, sp.FindFileURI("osrf://test_f2"));
+
+  // URI + env var
+  static char env[1024];
+  snprintf(env, sizeof(env), "%s", ("IGN_FILE_PATH=" + dir1).c_str());
+  putenv(env);
+
+  sp.SetFilePathEnv("IGN_FILE_PATH");
+  EXPECT_EQ("IGN_FILE_PATH", sp.FilePathEnv());
+  EXPECT_EQ(file1, sp.FindFileURI("anything://test_f1"));
+  EXPECT_NE(file2, sp.FindFileURI("anything://test_f2"));
+
+  std::string newEnv{"IGN_NEW_FILE_PATH"};
+  snprintf(env, sizeof(env), "%s", (newEnv + "=" + dir2).c_str());
+  putenv(env);
+
+  sp.SetFilePathEnv(newEnv);
+  EXPECT_EQ(newEnv, sp.FilePathEnv());
+  EXPECT_NE(file1, sp.FindFileURI("anything://test_f1"));
+  EXPECT_EQ(file2, sp.FindFileURI("anything://test_f2"));
 }
 
 //////////////////////////////////////////////////
@@ -310,7 +338,7 @@ TEST_F(SystemPathsFixture, FindFile)
   EXPECT_EQ("", sp.FindFile(this->filesystemRoot + "no_such_file"));
   EXPECT_EQ("", sp.FindFile("no_such_file"));
   EXPECT_EQ(file1, sp.FindFile(common::joinPaths("test_dir1", "test_f1")));
-  EXPECT_EQ(file1, sp.FindFile("file:test_dir1/test_f1"));
+  EXPECT_EQ(file1, sp.FindFile("file://test_dir1/test_f1"));
 
   // Existing absolute paths
 #ifndef _WIN32