+++ src/Util_TEST.cc
@@ -173,6 +173,46 @@ TEST(Util_TEST, emptyENV)
 }
 
 /////////////////////////////////////////////////
+TEST(Util_TEST, findFile)
+{
+  EXPECT_EQ("", ignition::common::findFile("no_such_file"));
+
+  // Create file to find
+  auto dir1 = ignition::common::absPath("test_dir1");
+  ignition::common::createDirectories(dir1);
+  auto file1 = ignition::common::absPath(
+      ignition::common::joinPaths(dir1, "test_f1"));
+
+  std::ofstream fout;
+  fout.open(file1, std::ofstream::out);
+  fout << "asdf";
+  fout.close();
+
+  // Relative path
+  auto systemPaths = ignition::common::systemPaths();
+  ASSERT_NE(nullptr, systemPaths);
+  EXPECT_EQ(file1, systemPaths->FindFile(common::joinPaths("test_dir1",
+      "test_f1"), true));
+
+  // Relative prefixed by file://
+  EXPECT_EQ(file1, ignition::common::findFile("file://test_dir1/test_f1"));
+
+  // Custom callback
+#ifndef _WIN32
+  const auto tmpDir = "/tmp";
+#else
+  const auto tmpDir = "C:\\Windows";
+#endif
+  auto fileCb = [&tmpDir](const ignition::common::URI &)
+  {
+    return tmpDir;
+  };
+
+  ignition::common::addFindFileURICallback(fileCb);
+  EXPECT_EQ(tmpDir, ignition::common::findFile("model://banana"));
+}
+
+/////////////////////////////////////////////////
 int main(int argc, char **argv)
 {
   ::testing::InitGoogleTest(&argc, argv);